# app/infonova-api/infonova-api-clients/swagger/tasksV2.json
{
  "swagger": "2.0",
  "info": {
    "title": "Infonova API Swagger",
    "version": ""
  },
  "host": "localhost:9700",
  "basePath": "/r6-api",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/{tenant}/tasks/v2/tasks/{id}": {
      "get": {
        "tags": [
          "Tasks v2"
        ],
        "summary": "GET /tasks/v2/tasks/{id}",
        "description": "GET /tasks/v2/tasks/{id}",
        "operationId": "getTasksV2TasksId",
        "produces": [
          "application/json",
          "application/vnd.infonova.r6+json"
        ],
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "$ref": "#/definitions/TaskV2DomainTask"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/CommonsDomainModelMessageAware"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/CommonsDomainModelMessageAware"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/CommonsDomainModelMessageAware"
            }
          }
        },
        "security": [
          {
            "oauth2ClientCredentialAuthentication": [
            ]
          },
          {
            "oauth2ResourceOwnerPasswordCredentialAuthentication": [
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "Tasks v2"
        ],
        "summary": "PUT /tasks/v2/tasks/{id}",
        "description": "PUT /tasks/v2/tasks/{id}",
        "operationId": "putTasksV2TasksId",
        "consumes": [
          "application/json",
          "application/vnd.infonova.r6+json"
        ],
        "produces": [
          "application/json",
          "application/vnd.infonova.r6+json"
        ],
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TaskV2DomainTask"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "$ref": "#/definitions/TaskV2DomainTask"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/CommonsDomainModelMessageAware"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/CommonsDomainModelMessageAware"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/CommonsDomainModelMessageAware"
            }
          }
        },
        "security": [
          {
            "oauth2ClientCredentialAuthentication": [
            ]
          },
          {
            "oauth2ResourceOwnerPasswordCredentialAuthentication": [
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "Tasks v2"
        ],
        "summary": "PATCH /tasks/v2/tasks/{id}",
        "description": "PATCH /tasks/v2/tasks/{id}",
        "operationId": "patchTasksV2TasksId",
        "consumes": [
          "application/json",
          "application/vnd.infonova.r6+json"
        ],
        "produces": [
          "application/json",
          "application/vnd.infonova.r6+json"
        ],
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TaskV2DomainTask"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "$ref": "#/definitions/TaskV2DomainTask"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/CommonsDomainModelMessageAware"
            }
          },
          "422": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/CommonsDomainModelMessageAware"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/CommonsDomainModelMessageAware"
            }
          }
        },
        "security": [
          {
            "oauth2ClientCredentialAuthentication": [
            ]
          },
          {
            "oauth2ResourceOwnerPasswordCredentialAuthentication": [
            ]
          }
        ]
      }
    },
...
  },
  "definitions": {
    "CommonsDomainModelArgument": {
      "type": "object",
      "properties": {
        "argument": {
          "type": "string"
        },
        "argumentDescription": {
          "type": "string"
        },
        "argumentKey": {
          "type": "string"
        },
        "subMessages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CommonsDomainModelMessage"
          }
        }
      }
    },
    "CommonsDomainModelLink": {
      "type": "object",
      "properties": {
        "href": {
          "type": "string"
        },
        "templated": {
          "type": "boolean"
        }
      }
    },
    "CommonsDomainModelMessage": {
      "type": "object",
      "properties": {
        "args": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CommonsDomainModelArgument"
          }
        },
        "key": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "messageTemplate": {
          "type": "string"
        }
      }
    },
    "CommonsDomainModelMessageAware": {
      "type": "object",
      "properties": {
        "logId": {
          "type": "string"
        },
        "messages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CommonsDomainModelMessage"
          }
        }
      }
    },
    "CommonsTimeInterval": {
      "type": "object",
      "properties": {
        "end": {
          "type": "string",
          "format": "date-time"
        },
        "start": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "PartyV1DomainContactMedium": {
      "type": "object",
      "properties": {
        "contactMedium": {
          "type": "string"
        },
        "contactType": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "parameters": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "postalAddress": {
          "type": "string"
        },
        "preferred": {
          "type": "boolean"
        },
        "role": {
          "type": "string"
        },
        "telephoneNumber": {
          "$ref": "#/definitions/PartyV1DomainTelephoneNumber"
        },
        "type": {
          "type": "string",
          "enum": [
            "Email",
            "TelephoneNumber",
            "PostalAddress",
            "Other"
          ]
        },
        "validFor": {
          "$ref": "#/definitions/CommonsTimeInterval"
        }
      }
    },
    "PartyV1DomainRelatedParty": {
      "type": "object",
      "properties": {
        "_links": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/CommonsDomainModelLink"
          }
        },
        "contactMedium": {
          "$ref": "#/definitions/PartyV1DomainContactMedium"
        },
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "party": {
          "type": "string",
          "format": "uuid"
        },
        "primary": {
          "type": "boolean"
        },
        "role": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "PartyV1DomainTelephoneNumber": {
      "type": "object",
      "properties": {
        "number": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "enum": [
            "Mobile",
            "FixedLine"
          ]
        }
      }
    },
    "ProductorderingV1DomainRelatedProductOrder": {
      "type": "object",
      "properties": {
        "_links": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/CommonsDomainModelLink"
          }
        },
        "id": {
          "type": "string"
        },
        "relationType": {
          "type": "string"
        },
        "tenant": {
          "type": "string"
        }
      }
    },
    "ResourceorderingV1DomainRelatedResourceOrder": {
      "type": "object",
      "properties": {
        "_links": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/CommonsDomainModelLink"
          }
        },
        "id": {
          "type": "string"
        },
        "relationType": {
          "type": "string"
        },
        "tenant": {
          "type": "string"
        }
      }
    },
    "ServiceorderingV1DomainRelatedServiceOrder": {
      "type": "object",
      "properties": {
        "_links": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/CommonsDomainModelLink"
          }
        },
        "id": {
          "type": "string"
        },
        "relationType": {
          "type": "string"
        },
        "tenant": {
          "type": "string"
        }
      }
    },
...
    "TaskV2DomainTask": {
      "type": "object",
      "properties": {
        "_links": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/CommonsDomainModelLink"
          }
        },
        "assignee": {
          "type": "string"
        },
        "billingAccount": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "creationDate": {
          "type": "string",
          "format": "date-time"
        },
        "creator": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "parameters": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "priority": {
          "type": "integer",
          "format": "int32"
        },
        "relatedParties": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PartyV1DomainRelatedParty"
          }
        },
        "relatedProductOrders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductorderingV1DomainRelatedProductOrder"
          }
        },
        "relatedResourceOrders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ResourceorderingV1DomainRelatedResourceOrder"
          }
        },
        "relatedServiceOrders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ServiceorderingV1DomainRelatedServiceOrder"
          }
        },
        "resolutionAction": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "taskDefinition": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "oauth2ClientCredentialAuthentication": {
      "type": "oauth2",
      "description": "Client credentials flow configuration for the Infonova Keycloak.",
      "flow": "application",
      "tokenUrl": ""
    },
    "oauth2ResourceOwnerPasswordCredentialAuthentication": {
      "type": "oauth2",
      "description": "Resource Owner Password Credentials flow configuration for the Infonova Keycloak.",
      "flow": "password",
      "tokenUrl": ""
    }
  },
...
}